requirements.txt
requests
google-generativeai
google-adk
python-dotenv  